actrl.py: Calculate off threshold from min power time
Rather than overall on time. The goal is to use the initial power
threshold when reducing to minimum power, to be more tolerant of
overshoot and try and stay in min power rather than turning off.

The 90 min purge cycle will be out of sync in this case, but that
is already likely after extended runtime.

Should not change behaviour in the case that the system is running
with demand slightly below min power for an extended time.